Split docker component into several others to integrate it in config
