+++ src/app/components/dockeredit/dockeredit.component.html
@@ -1,0 +1,52 @@
+<div class="modal-dialog">
+    <div class="modal-content">
+        <div class="modal-header">
+            <h4 class="modal-title">Edit instance</h4>
+        </div>
+        <div class="modal-body">
+            <form>
+                <div class="input-group mb-3">
+                    <div class="input-group-prepend">
+                        <label for="idField" class="input-group-text">Id</label>
+                    </div>
+                    <input id="idField" class="form-control" type="text" name="id" [(ngModel)]="id" readonly="true">
+
+                </div>
+
+                <div class="input-group mb-3">
+                    <div class="input-group-prepend">
+                        <label for="hostField" class="input-group-text">Host</label>
+                    </div>
+                    <input id="hostField" class="form-control" type="text" name="host" (input)="modified=true" [(ngModel)]="host">
+                </div>
+
+                <div class="input-group mb-3">
+                    <div class="input-group-prepend">
+                        <label for="aliasField" class="input-group-text">Alias</label>
+                    </div>
+                    <input id="aliasField" class="form-control" type="text" name="alias" (input)="modified=true" [(ngModel)]="alias">
+
+                </div>
+
+                <div class="input-group mb-3">
+                    <div class="input-group-prepend">
+                        <label for="idField" class="input-group-text">Connected</label>
+                    </div>
+                    <input id="idField" class="form-control" type="text" name="connected" [(ngModel)]="connected" readonly="true">
+
+                </div>
+            </form>
+            <app-dockerhandshake *ngIf="handshake !== null" [handshake]="handshake"></app-dockerhandshake>
+
+        </div>
+        <div class="modal-footer">
+            <div class="input-group mb-3">
+                <button class="btn btn-success mr-2" (click)="updateDockerInstance()" [disabled]="!modified || updateLock || handshake !== null">Save</button>
+                <button class="btn btn-primary mr-2" (click)="emitDone()" [disabled]="updateLock || handshake !== null">Cancel</button>
+                <button class="btn btn-primary mr-2" (click)="testConnection();" [disabled]="updateLock || handshake !== null">Test Connectivity</button>
+                <button *ngIf="!connected && handshake === null" class="btn btn-primary mr-2" (click)="reconnectToDockerInstance()" [disabled]="updateLock || handshake !== null">Reconnect</button>
+                <button class="btn btn-danger mr-2" (click)="removeDockerInstance();" [disabled]="lock || handshake !== null">Delete</button>
+            </div>
+        </div>
+    </div>
+</div>
