+++ src/app/containers/default-layout/default-layout.component.html
@@ -52,12 +52,6 @@
             </div>
         </li>
 
-	<li class="nav-item px-3">
-	    <a [routerLink]="['/views/docker']" class="nav-link" routerLinkActive="active">
-		<i class="fa fa-cloud"></i><span class="d-md-down-none">Docker</span>
-	    </a>
-	</li>
-
         <li class="nav-item px-3">
             <a [routerLink]="['/views/adapters']" class="nav-link" routerLinkActive="active">
                 <i class="fa fa-database"></i><span class="d-md-down-none">Adapters</span>
