--- src/app/views/docker/docker.component.html
@@ -1,136 +1,0 @@
-<app-toast></app-toast>
-
-<div *ngIf="notfound">
-    No docker instance with that id.<br>
-    <button class="btn btn-primary" routerLink="..">Go Back</button>
-</div>
-
-<div *ngIf="!notfound">
-    <div *ngIf="id===-1">
-        <form (ngSubmit)="addDockerInstance()">
-            <div class="input-group mb-3">
-                <div class="input-group-prepend">
-                    <label for="hostField" class="input-group-text">Host</label>
-                </div>
-                <input id="hostField" class="form-control" type="text" name="host" (input)="hostInput()" [(ngModel)]="host" [disabled]="handshake!==null" required>
-            </div>
-
-            <div class="input-group mb-3">
-                <div class="input-group-prepend">
-                    <label for="aliasField" class="input-group-text">Alias</label>
-                </div>
-                <input id="aliasField" class="form-control" type="text" name="alias" (input)="aliasInput()" [(ngModel)]="alias" [disabled]="handshake!==null" required>
-
-            </div>
-
-            <div class="input-group mb-3">
-                <input class="btn btn-success mr-2" type="submit" value="Add" [disabled]="handshake!==null">
-                <button class="btn btn-secondary" routerLink=".." [disabled]="handshake!==null && handshake.status!=='NOT_RUNNING' && handshake.status!=='FAILED'">Cancel</button>
-            </div>
-
-        </form>
-        <div *ngIf="dockerSetupResult!==null&&dockerSetupResult.error!==''" class="alert alert-danger">
-            {{ dockerSetupResult.error }}
-        </div>
-
-
-    </div>
-    <div *ngIf="id!==-1">
-        <form>
-            <div class="input-group mb-3">
-                <div class="input-group-prepend">
-                    <label for="idField" class="input-group-text">Id</label>
-                </div>
-                <input id="idField" class="form-control" type="text" name="id" [(ngModel)]="id" readonly="true">
-
-            </div>
-
-            <div class="input-group mb-3">
-                <div class="input-group-prepend">
-                    <label for="hostField" class="input-group-text">Host</label>
-                </div>
-                <input id="hostField" class="form-control" type="text" name="host" (input)="modified=true" [(ngModel)]="host">
-            </div>
-
-            <div class="input-group mb-3">
-                <div class="input-group-prepend">
-                    <label for="aliasField" class="input-group-text">Alias</label>
-                </div>
-                <input id="aliasField" class="form-control" type="text" name="alias" (input)="modified=true" [(ngModel)]="alias">
-
-            </div>
-
-            <div class="input-group mb-3">
-                <div class="input-group-prepend">
-                    <label for="idField" class="input-group-text">Connected</label>
-                </div>
-                <input id="idField" class="form-control" type="text" name="connected" [(ngModel)]="connected" readonly="true">
-
-            </div>
-
-            <div class="input-group mb-3">
-                <button class="btn btn-success mr-2" (click)="updateDockerInstance()" [disabled]="!modified||updateLock||handshake!==null">Save</button>
-                <button class="btn btn-primary mr-2" routerLink=".." [disabled]="updateLock||handshake!==null">Cancel</button>
-                <button class="btn btn-primary mr-2" (click)="testConnection();" [disabled]="updateLock||handshake!==null">Test Connectivity</button>
-                <button *ngIf="!connected&&handshake===null" class="btn btn-primary mr-2" (click)="reconnectToDockerInstance()" [disabled]="updateLock||handshake!==null">Reconnect</button>
-                <button class="btn btn-danger mr-2" (click)="removeDockerInstance();" [disabled]="lock||handshake!==null">Delete</button>
-            </div>
-
-
-        </form>
-    </div>
-
-    <div *ngIf="handshake!==null">
-        <div *ngIf="handshake.status=='RUNNING'||handshake.status=='NOT_RUNNING'">
-            Run the following command on {{ handshake.hostname }}:<br>
-            <div *ngIf="handshake.containerExists==='true'">
-                <code>
-                    {{ handshake.execCommand }}
-                </code>
-                <br>
-                <button class="btn btn-secondary" (click)="_util.clipboard(handshake.execCommand)">Copy</button>
-                <details>
-                    <summary>No docker container running?</summary>
-                    Use this command:<br>
-                    <code>
-                        {{ handshake.runCommand }}
-                    </code>
-                    <br>
-                    <button class="btn btn-secondary mb-2" (click)="_util.clipboard(handshake.runCommand)">Copy</button>
-                </details>
-            </div>
-            <div *ngIf="handshake.containerExists==='false'">
-                <code>
-                    {{ handshake.runCommand }}
-                </code>
-                <br>
-                <button class="btn btn-secondary" (click)="_util.clipboard(handshake.runCommand)">Copy</button>
-                <details>
-                    <summary>Is there a docker container already running?</summary>
-                    Use this command:<br>
-                    <code>
-                        {{ handshake.execCommand }}
-                    </code>
-                    <br>
-                    <button class="btn btn-secondary mb-2" (click)="_util.clipboard(handshake.execCommand)">Copy</button>
-                </details>
-            </div>
-        </div>
-
-        <div *ngIf="handshake.status=='RUNNING'">
-            <button class="btn btn-danger" (click)="cancelHandshake()">Cancel handshake</button>
-        </div>
-
-        <div *ngIf="handshake.status==='NOT_RUNNING' || handshake.status==='FAILED'">
-            <div *ngIf="handshake.lastErrorMessage!=''" class="alert alert-danger">
-                Error: {{ handshake.lastErrorMessage }}
-            </div>
-            <button class="btn btn-primary" (click)="redoHandshake()">Try again</button>
-        </div>
-
-        <div *ngIf="handshake.status=='SUCCESS'" class="alert alert-success">
-            Successful handshake at {{ handshake.lastErrorMessage }}.
-        </div>
-
-    </div>
-</div>
