+++ src/app/views/dockerconfig/dockerconfig.component.html
@@ -2,28 +2,34 @@
 
 <div class="container">
     <div class="row justify-content-start">
-	<div *ngFor="let instance of instances" class="col-4">
-	    <div class="card">
-		<div class="card-header">{{ instance.alias }}</div>
-		<div class="card-body">
-		    Host: {{ instance.host }}<br>
-		    Connected: {{ instance.connected }}<br>
-		    <a class="btn btn-primary mr-2" [routerLink]="['.', instance.id]">Edit</a>
-		    <button class="btn btn-danger mr-2" (click)="removeDockerInstance(instance);">Delete</button>
-		</div>
-	    </div>
-	</div>
+        <div *ngFor="let instance of instances" class="col-4">
+            <div class="card">
+                <div class="card-header">{{ instance.alias }}</div>
+                <div class="card-body">
+                    Host: {{ instance.host }}<br>
+                    Connected: {{ instance.connected }}<br>
+                    <button class="btn btn-primary mr-2" (click)="showModal(instance.id)">Edit</button>
+                    <button class="btn btn-danger mr-2" (click)="removeDockerInstance(instance);">Delete</button>
+                </div>
+            </div>
+        </div>
     </div>
     <div class="row justify-content-start">
         <div class="col">
-	    <a class="btn btn-success" [routerLink]="['.', 'new']">Add new instance</a>
+            <!--<a class="btn btn-success" [routerLink]="['.', 'new']">Add new instance</a>-->
+            <button class="btn btn-primary" (click)="showModal(null)">Add new instance</button>
         </div>
-        <div *ngIf="status.available&&(!status.connected||autoConnectRunning)" class="col">
-	    <button class="btn btn-success" (click)="autoDocker()" [disabled]="status.running||autoConnectRunning">Autoconnect to localhost</button>
+        <div *ngIf="status.available && (!status.connected || autoConnectRunning)" class="col">
+            <button class="btn btn-success" (click)="autoDocker()" [disabled]="status.running || autoConnectRunning">Autoconnect to localhost</button>
         </div>
     </div>
 </div>
 
 <div *ngIf="error!==null" class="alert alert-danger">
     Cannot load list of docker instances: {{ error }}
+</div>
+
+<div bsModal #dockerConfigModal="bs-modal" id="addNewInstance" class="modal fade" data-bs-backdrop="static" data-bs-keyboard="false" aria-labelledby="Add/Edit a docker instance" aria-hidden="true">
+    <app-dockernew *ngIf="modalId === null" (done)="closeModal()"></app-dockernew>
+    <app-dockeredit *ngIf="modalId !== null" [id]="modalId" (done)="closeModal()"></app-dockeredit>
 </div>
