+++ src/app/views/views-routing.module.ts
@@ -10,7 +10,6 @@ import {QueryingComponent} from './querying/querying.component';
 import {AdaptersComponent} from './adapters/adapters.component';
 import {AboutComponent} from './about/about.component';
 import {QueryInterfacesComponent} from './query-interfaces/query-interfaces.component';
-import {DockerComponent} from './docker/docker.component';
 import {DockerconfigComponent} from './dockerconfig/dockerconfig.component';
 
 const routes: Routes = [
@@ -119,24 +118,24 @@ const routes: Routes = [
         }
     },
     {
+        path: 'config/dockerConfig',
+        component: DockerconfigComponent,
+        data: {
+            title: 'Docker Setup'
+        }
+    },
+    {
+        path: 'config/dockerPage',
+        component: DockerconfigComponent,
+        data: {
+            title: 'Docker Setup'
+        }
+    },
+    {
         path: 'config/:page',
         component: FormGeneratorComponent,
         data: {
             title: 'Form Generator'
-        }
-    },
-    {
-        path: 'docker',
-        component: DockerconfigComponent,
-        data: {
-            title: 'Docker'
-        }
-    },
-    {
-        path: 'docker/:id',
-        component: DockerComponent,
-        data: {
-            title: 'Docker'
         }
     },
     {
